added user auth and route protection

+++ frontend/workout-app/src/components/WorkoutDetails.js
@@ -1,17 +1,31 @@
 import { Link } from "react-router-dom";
 import useWorkoutContext from "../hooks/useWorkoutContext";
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
+import useAuthContext from "../hooks/userAuthContext";
+
 
 const WorkoutDetails = ({ workout }) => {
     const { dispatch } = useWorkoutContext()
+    const { user } = useAuthContext()
+
     const handleClick = async() => {
+        
+        if (!user) {
+            console.log('not logged in')
+            return
+        }
+
         const response = await fetch("http://localhost:4000/api/workouts/" + workout._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         })
-
+        
         const json = await response.json()
-
+        
         if (response.ok) {
+            
             dispatch({type: 'DELETE_WORKOUT', payload: json})
         }
     }
